Clean up util file

+++ src/grower/util.cpp
@@ -12,7 +12,7 @@ const char * Log::s_log_filename =
     getenv_default("POMAGMA_LOG_FILE", DEFAULT_LOG_FILE);
 std::ofstream Log::s_log_stream(s_log_filename, std::ios_base::app);
 
-const unsigned Log::s_log_level(
+const size_t Log::s_log_level(
     getenv_default("POMAGMA_LOG_LEVEL", DEFAULT_LOG_LEVEL));
 
 //----------------------------------------------------------------------------
